in the current mapping file there are no UUIDs for the
source flows (but for the target flows) and only the
sub-compartment names provided (e.g. `air` which could
be `emission > air` or `resource > air`)
(https://github.com/USEPA/Federal-LCA-Commons-Elementary-Flow-List/blob/flowmappingforolca/fedelemflowlist/flowmapping/FedElemFlowList_0.1_mapping.csv)

also, we only have the flow property name of the source
flow and the unit of the target flow in the mapping.

we try to generate the UUIDs like in the ElectricityLCI
project (which is our current test database):
https://github.com/USEPA/ElectricityLCI/blob/master/electricitylci/olca_jsonld_writer.py#L142

However, this is not unambiguous, e.g.:

```python
def _catpath(*args) -> str:
    p = ''
    for arg in args:
        if _isnil(arg):
            continue
        if p != '':
            p = p + "/"
        p = p + str(arg).strip()

    if p == 'air':
        return 'Elementary flows/emission/air'
    if p == 'ground':
        return 'Elementary flows/resource/ground'
    if p == 'soil':
        return 'Elementary flows/emission/soil'
    if p == 'water':
        return 'Elementary flows/emission/water'
    return p
```

... we need to check the current workarounds when the
mapping is more complete.